refactor: separated user profile and edit profile
