+++ client/components/index.js
@@ -7,6 +7,7 @@ export {default as Navbar} from './navbar'
 export {default as Footer} from './footer'
 export {default as Home} from './home'
 export {default as UserProfile} from './user-profile'
+export {default as UserProfileEdit} from './user-profile-edit'
 export {default as LogIn} from './log-in-auth-form'
 export {default as SignUp} from './sign-up-auth-form'
 export {default as RecordingScreen} from './recording-screen'
